Rename voyager1_pm to voyager1_depart_target.py since it now does Centaur burn 2 also.
